Move menus in to dedicated module
It was really confusing having the menu's and the options mixed
together. This is a much better situation.
