+++ HarpGuru/OverlayMenu/Menus/LayoutMenu/types/index.ts
@@ -1,8 +1,8 @@
 import type { HarpStrata } from 'harpstrata'
 
-import {DisplayModes} from '../../../HarpFace'
+import type { DisplayModes } from '../../../../HarpFace'
 
-export type CovariantMenuProps = {
+export type LayoutMenuProps = {
   readonly activeHarpStrata: HarpStrata,
   readonly setActiveHarpStrata: (arg0: HarpStrata) => void
   readonly activeDisplayMode: DisplayModes
