+++ HarpGuru/OverlayMenu/Menus/MenuContainer/index.tsx
@@ -1,7 +1,7 @@
 import { View } from 'react-native'
 import React, { ReactNode, ReactElement } from 'react'
 
-import { styles } from '../styles'
+import { styles } from '../../styles'
 
 export const MenuContainer = ({children}: {readonly children: ReactNode}): ReactElement => {
   return (
